# routes/users.js
const express = require('express');
const router = express.Router();
const User = require('../controllers/User');
const {checkUser, checkUserWithoutPhone, checkAdmin} = require('../libs/jwt');
// const passport = require('passport');
// require('../libs/facebookAuth');

// router.get('/login/facebook', passport.authenticate('facebook', { scope : 'email' }));

// router.get('/login/facebook/callback', async (req, res, next) => {
//     try {
//         const result = await User.facebookLogin(req.query);
//         res.send(result);
//     } catch (err) {
//         next(err);
//     }
// });

// router.post('/register', async (req, res, next) => {
//     try {
//         const result = await User.register(req.body);
//         res.send(result);
//     } catch (err) {
//         next(err);
//     }
// });

// router.put('/login/facebook/phone', checkUserWithoutPhone, async (req, res, next) => {
//     try {
//         const result = await User.facebookLoginPhoneUpdate(req.user, req.body);
//         res.send(result);
//     } catch (err) {
//         next(err);
//     }
// });

// router.put('/login/facebook/phone_confirm', checkUserWithoutPhone, async (req, res, next) => {
//     try {
//         const result = await User.facebookLoginPhoneConfirm(req.user, req.body);
//         res.send(result);
//     } catch (err) {
//         next(err);
//     }
// });

router.get('/', checkAdmin, async (req, res, next) => {
    try {
        const {page, limit, phone, firstName, lastName, middleName, type} = req.query;
        const result = await User.getUserList(page, limit, phone, firstName, lastName, middleName, type);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.get('/:id', checkAdmin, async (req, res, next) => {
    try {
        const result = await User.getUserById(req.params.id);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.put('/check_code', async (req, res, next) => {
    try {
        const {phone, code} = req.body;
        const result = await User.checkCode(phone, code);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.post('/coords', async (req, res, next) => {
    try {
        const {lat, lon} = req.body;
        const result = await User.updateUserCoords(req.user, lat, lon);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.post('/problem_request', async (req, res, next) => {
    try {
        const {description, scooterId} = req.body;
        const result = await User.createProblemRequest(req.user._id, description, scooterId);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

// router.post('/login', async (req, res, next) => {
//     try {
//         const result = await User.loginAndSendCode(req.body);
//         res.send(result);
//     } catch (err) {
//         next(err);
//     }
// });

router.post('/firebase_id', checkUser, async (req, res, next) => {
    try {
        const {firebaseId} = req.body;
        const result = await User.putFirebaseId(req.user, firebaseId);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.delete('/firebase_id', checkUser, async (req, res, next) => {
    try {
        const {firebaseId} = req.body;
        const result = await User.deleteFirebaseId(req.user, firebaseId);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.post('/code/confirm', async (req, res, next) => {
    try {
        const {phone, code} = req.body;
        const result = await User.confirmCode(phone,code);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.post('/code', async (req, res, next) => {
    try {
        const {phone} = req.body;
        const result = await User.sendCode(phone);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.put('/reset_password', async (req, res, next) => {
    try {
        const result = await User.resetPassword(req.body);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

router.get('/info', checkUser, async (req, res, next) => {
    try {
        const fields = ['fb', '_id', 'phone', 'balance'];
        const result = {};
        for (const f of fields)
            result[f] = req.user[f];
        res.send({user: {...req.user, ...result}});
    } catch (err) {
        next(err);
    }
});

router.put('/info', checkUser, async (req, res, next) => {
    try {
        const {firstName, lastName, middleName, email, birthday} = req.body;
        const result = await User.updateInfo(req.user, firstName, lastName, middleName, email, birthday);
        res.send(result);
    } catch (err) {
        next(err);
    }
});

module.exports = router;