# swagger.yaml
swagger: '2.0'
info:
  title: scooter
  version: 1.0.0
host: 'school.vvdev.ru'
schemes:
  - https
basePath: /api
produces:
  - application/json
paths:
...
  /users:
    get:
      tags:
      - Admin
      summary: user list
      parameters:
      - in: query
        name: limit
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: firstName
        type: string
      - in: query
        name: lastName
        type: string
      - in: query
        name: middleName
        type: string
      - in: query
        name: phone
        type: string
      - in: query
        name: type
        type: string
      responses:
        200:
          description: ok
  /users/{id}:
    get:
      tags:
      - Admin
      summary: user list
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok        
...
  /users/code:
    post:
      tags:
      - Users
      summary: Send code method for register of user
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            phone:
              type: string
              description: 'format +380XXXXXXXXX'
      responses:
        200:
          description: Send code on phone number to user
          schema:
            type: object
            properties:
              message:
                type: string
              code:
                type: string
                description: Code sent on client only in test-version
        400:
          description: Incorrect phone number
        422:
          description: too many requests on send code(only one request in two minutes)
        500:
          description: server error
  /users/firebase_id:
    post:
      tags:
      - Users
      summary: добавить firebaseId
      parameters:
      - in: body
        name: bod
        schema:
          type: object
          properties:
            firebaseId:
              type: string
      responses:
        200:
          description: ok
    delete:
      tags:
      - Users
      summary: удалить firebaseId
      parameters:
      - in: body
        name: bod
        schema:
          type: object
          properties:
            firebaseId:
              type: string
      responses:
        200:
          description: ok
  /users/check_code:
    put:
      tags:
      - Users
      summary: Check code before register or reset password
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            phone:
              type: string
            code:
              type: string
      responses:
        200:
          description: OK
        400:
          description: Enter data
        401:
          description: Auth error
        500:
          description: Server error
  /users/code/confirm:
    post:
      tags:
      - Users
      summary: enter code to confirm login user
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            phone:
              type: string
            code:
              type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/token'
        400:
          description: Enter phone/code
        401:
          description: Auth error
        500:
          description: server error
  /users/info:
    get:
      tags:
      - Users
      summary: Get private information
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/user'
        500:
          description: server error
    put:
      tags:
      - Users
      summary: update private information
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            middleName:
              type: string
            email:
              type: string
            birthday:
              type: string
              description: ISOstring
      responses:
        200:
          description: OK
        500:
          description: server error
  /users/coords:
    post:
      tags:
      - Users
      summary: update coords of user
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            lat:
              type: number
            lon:
              type: number
      responses:
        200:
          description: OK
        500:
          description: server error
...
definitions:
...
  token:
    type: object
    properties:
      token:
        type: string
        description: 'send this token in headers x-access-token' 
      _id:
        type: string
        description: user ID
  user:
    type: object
    properties:
      _id: 
        type: string
      phone:
        type: string
      balance:
        type: number